requirements.txt
conda-forge::datasets
conda-forge::pandas
conda-forge::transformers
conda-forge::evaluate
conda-forge::wandb
conda-forge::accelerate

pytorch::pytorch
pytorch::torchvision
